Redirecting now works correctly.
