+++ classes/class06/twoter/app.js
@@ -12,32 +12,7 @@ var mongoose = require('mongoose');
 var mongoURI = process.env.MONGOURI || "mongodb://localhost/test";
 mongoose.connect(mongoURI);
 
-app = express();
-
-var config = require('./oauth.js');
-var FacebookStrategy = require('passport-facebook').Strategy;
-
-// Serialize and deserialize.
-passport.serializeUser(function(user, done) {
-done(null, user);
-});
-passport.deserializeUser(function(obj, done) {
-done(null, obj);
-});
-
-// Config
-passport.use(new FacebookStrategy({
- clientID: config.facebook.clientID,
- clientSecret: config.facebook.clientSecret,
- callbackURL: config.facebook.callbackURL
-},
-function (accessToken, refreshToken, profile, done) {
- process.nextTick(function () {
-   return done(null, profile);
- });
-}
-));
-
+require("./authentication");
 app = express();
 
 // Template engine.
